+++ A12_SpatialOpt/SpatialTree.cpp
@@ -33,15 +33,15 @@ void Node::split(int level) {
 
 void Node::checkCollisions() {
 	if ((children.size() > 0)) {
-		for (int i = 0; i < children.size(); i++)
+		for (int i = 0; i < (int)children.size(); i++)
 			children[i]->checkCollisions();
 	}
 	else if (objects.size() > 0) {
 		//check collisions between all the objects
-		for (int i = 0; i < objects.size() - 1; i++) {
+		for (int i = 0; i < (int)objects.size() - 1; i++) {
 			//if (!(objects[i]->checked)) {
 			//	objects[i]->checked = true;
-			for (int j = i + 1; j < objects.size(); j++) {
+			for (int j = i + 1; j < (int)objects.size(); j++) {
 				if (objects[i]->IsColliding(objects[j])) {
 					objects[i]->colliding = true;
 					objects[j]->colliding = true;
@@ -71,7 +71,7 @@ bool Node::containsObject(MyBOClass* object) {
 
 void Node::addObject(MyBOClass* object) {
 	if (children.size() > 0) {
-		for (int i = 0; i < children.size(); i++) {
+		for (int i = 0; i < (int)children.size(); i++) {
 			if (children[i]->containsObject(object)) {
 				MyBOClass* newObject = object;
 				children[i]->addObject(newObject);
@@ -88,8 +88,8 @@ void Node::addObject(MyBOClass* object) {
 void SpatialTree::checkCollisions() {
 	if (optimize) head->checkCollisions();
 	else {
-		for (int i = 0; i < objects.size(); i++) {
-			for (int j = 0; j < objects.size(); j++) {
+		for (int i = 0; i < (int)objects.size(); i++) {
+			for (int j = 0; j < (int)objects.size(); j++) {
 				if (i == j) continue;
 				if (objects[i]->IsColliding(objects[j]))
 					objects[i]->colliding = true;
@@ -124,10 +124,10 @@ void SpatialTree::addObject(MyBOClass * object) {
 
 void SpatialTree::displayTree(MeshManagerSingleton* Mesh) {
 	if (!display) return;
-	for (int i = 0; i < head->children.size(); i++) {
+	for (int i = 0; i < (int)head->children.size(); i++) {
 		Node* curr = head->children[i];
 		Mesh->AddCubeToRenderList(IDENTITY_M4 * glm::translate(curr->position.x, curr->position.y, curr->position.z) * glm::scale(curr->widths.x, curr->widths.y, curr->widths.z), RERED, WIRE);
-		for (int j = 0; j < curr->children.size(); j++) {
+		for (int j = 0; j < (int)curr->children.size(); j++) {
 			Node* curr2 = curr->children[j];
 			displayTree(Mesh, curr2);
 		}
@@ -139,10 +139,10 @@ void SpatialTree::displayTree(MeshManagerSingleton* Mesh, Node* node) {
 		Mesh->AddCubeToRenderList(IDENTITY_M4 * glm::translate(node->position.x, node->position.y, node->position.z) * glm::scale(node->widths.x, node->widths.y, node->widths.z), RERED, WIRE);
 		return;
 	}
-	for (int i = 0; i < node->children.size(); i++) {
+	for (int i = 0; i < (int)node->children.size(); i++) {
 		Node* curr = node->children[i];
 		Mesh->AddCubeToRenderList(IDENTITY_M4 * glm::translate(curr->position.x, curr->position.y, curr->position.z) * glm::scale(curr->widths.x, curr->widths.y, curr->widths.z), RERED, WIRE);
-		for (int j = 0; j < curr->children.size(); j++) {
+		for (int j = 0; j < (int)curr->children.size(); j++) {
 			Node* curr2 = curr->children[j];
 			displayTree(Mesh, curr2);
 		}
@@ -157,7 +157,7 @@ void SpatialTree::generateTree(int depth) {
 	head->position = m_v3Min + head->widths / 2.0f;
 
 	head->split(depth);
-	for (int i = 0; i < objects.size(); i++) {
+	for (int i = 0; i < (int)objects.size(); i++) {
 		head->addObject(objects[i]);
 	}
 }
